👌 Using single handler for all checkbox events. Moving logic to the Markdown component
