+++ packages/design-system/src/components/N8nMarkdown/Markdown.vue
@@ -6,6 +6,7 @@
 			:class="$style[theme]"
 			@click="onClick"
 			@mousedown="onMouseDown"
+			@change="onChange"
 			v-html="htmlContent"
 		/>
 		<div v-else :class="$style.markdown">
@@ -18,7 +19,7 @@
 </template>
 
 <script lang="ts" setup>
-import { computed } from 'vue';
+import { computed, ref } from 'vue';
 import type { Options as MarkdownOptions } from 'markdown-it';
 import Markdown from 'markdown-it';
 import markdownLink from 'markdown-it-link-attributes';
@@ -27,7 +28,7 @@ import markdownTaskLists from 'markdown-it-task-lists';
 import xss, { friendlyAttrValue, whiteList } from 'xss';
 
 import N8nLoading from '../N8nLoading';
-import { escapeMarkdown } from '../../utils/markdown';
+import { escapeMarkdown, replaceNth } from '../../utils/markdown';
 
 interface IImage {
 	id: string;
@@ -79,6 +80,8 @@ const props = withDefaults(defineProps<MarkdownProps>(), {
 		},
 	}),
 });
+
+const editor = ref<HTMLDivElement | undefined>(undefined);
 
 const { options } = props;
 const md = new Markdown(options.markdown)
@@ -149,7 +152,7 @@ const htmlContent = computed(() => {
 	return safeHtml;
 });
 
-const $emit = defineEmits(['markdown-click']);
+const $emit = defineEmits(['markdown-click', 'update-content']);
 const onClick = (event: MouseEvent) => {
 	let clickedLink: HTMLAnchorElement | null = null;
 
@@ -166,12 +169,42 @@ const onClick = (event: MouseEvent) => {
 	$emit('markdown-click', clickedLink, event);
 };
 
+// Handle checkbox changes
+const onChange = async (event: Event) => {
+	if (event.target instanceof HTMLInputElement) {
+		const checkboxes = editor.value?.querySelectorAll('input[type="checkbox"]');
+		if (checkboxes) {
+			// Get the index of the checkbox that was clicked
+			const index = Array.from(checkboxes).indexOf(event.target);
+			if (index !== -1) {
+				await onCheckboxChange(event.target, index);
+			}
+		}
+	}
+};
+
 const onMouseDown = (event: MouseEvent) => {
 	// Mouse down on input fields is caught by node view handlers
 	// which prevents checking them, this will prevent that
 	if (event.target instanceof HTMLInputElement) {
 		event.stopPropagation();
 	}
+};
+
+// Update markdown when checkbox state changes
+const onCheckboxChange = async (checkbox: HTMLInputElement, position: number) => {
+	const currentContent = props.content;
+	if (!currentContent) {
+		return;
+	}
+	// We are using position to connect the checkbox with the corresponding line in the markdown
+	const newContent = replaceNth(
+		currentContent,
+		/\- \[[x|\s]\]/gim,
+		checkbox.checked ? '- [x]' : '- [ ]',
+		position + 1,
+	);
+	$emit('update-content', newContent);
 };
 </script>
