+++ packages/design-system/src/components/N8nSticky/Sticky.vue
@@ -32,6 +32,7 @@
 					:content="modelValue"
 					:with-multi-breaks="true"
 					@markdown-click="onMarkdownClick"
+					@update-content="$emit('update:modelValue', $event)"
 				/>
 			</div>
 			<div
@@ -69,7 +70,6 @@ import N8nMarkdown from '../N8nMarkdown';
 import N8nResizeWrapper, { type ResizeData } from '../N8nResizeWrapper/ResizeWrapper.vue';
 import N8nText from '../N8nText';
 import { useI18n } from '../../composables/useI18n';
-import { replaceNth } from '../../utils/markdown';
 
 interface StickyProps {
 	modelValue?: string;
@@ -111,8 +111,6 @@ const $emit = defineEmits<{
 const { t } = useI18n();
 const isResizing = ref(false);
 const input = ref<HTMLTextAreaElement | undefined>(undefined);
-const markdownContainer = ref<HTMLDivElement | undefined>(undefined);
-const checkboxListeners = ref<Map<string, EventListener>>(new Map());
 
 const resHeight = computed((): number => {
 	return props.height < props.minHeight ? props.minHeight : props.height;
@@ -146,14 +144,6 @@ watch(
 		}, 100);
 	},
 );
-
-onMounted(() => {
-	initCheckboxEvents();
-});
-
-onBeforeUnmount(() => {
-	removeCheckboxListeners();
-});
 
 const onDoubleClick = () => {
 	if (!props.readOnly) $emit('edit', true);
@@ -190,56 +180,6 @@ const onInputScroll = (event: WheelEvent) => {
 	if (!event.ctrlKey && !event.metaKey) {
 		event.stopPropagation();
 	}
-};
-
-// TODO: Ideally, this should be part of the Markdown component
-const initCheckboxEvents = () => {
-	const markdownContainerEl = markdownContainer.value;
-	const checkboxes = markdownContainerEl?.querySelectorAll('input[type="checkbox"]');
-	if (checkboxes) {
-		checkboxes.forEach((checkbox, position) => {
-			const handler: EventListener = onCheckboxChange.bind(null, checkbox, position);
-			checkbox.addEventListener('change', handler);
-			checkboxListeners.value.set(checkbox.id, handler);
-		});
-	}
-};
-
-const removeCheckboxListeners = () => {
-	const markdownContainerEl = markdownContainer.value;
-	const checkboxes = markdownContainerEl?.querySelectorAll('input[type="checkbox"]');
-	if (checkboxes) {
-		checkboxes.forEach((checkbox) => {
-			const handler = checkboxListeners.value.get(checkbox.id);
-			if (handler) {
-				checkbox.removeEventListener('change', handler);
-				checkboxListeners.value.delete(checkbox.id);
-			}
-		});
-	}
-};
-
-const resetCheckboxListeners = async () => {
-	removeCheckboxListeners();
-	await nextTick();
-	initCheckboxEvents();
-};
-
-// Update markdown when checkbox state changes
-const onCheckboxChange = async (checkbox: HTMLInputElement, position: number) => {
-	const currentContent = props.modelValue;
-	if (!currentContent) {
-		return;
-	}
-	// We are using position to connect the checkbox with the corresponding line in the markdown
-	const newContent = replaceNth(
-		currentContent,
-		/\- \[[x|\s]\]/gim,
-		checkbox.checked ? '- [x]' : '- [ ]',
-		position + 1,
-	);
-	$emit('update:modelValue', newContent);
-	await resetCheckboxListeners();
 };
 </script>
